terminada aplicacion, ahora en desktop. Configuracion Electron ok

+++ client/src/components/AppointmentDetail.js
@@ -1,6 +1,7 @@
 import React, { Fragment } from 'react'
 import { Link, withRouter } from 'react-router-dom'
 import axiosClient from '../config/axios'
+import Swal from 'sweetalert2'
 
 //añado withRouter para tener las props por default y le añado props como parámtetro para poder acceder a history.push()
 const AppointmentDetail = (props) => {
@@ -16,17 +17,40 @@ const AppointmentDetail = (props) => {
     //cuando elimino, redirecciona correctamente pero para ver que lo ha eliminado tengo que recargar. Para que esto no pase, utilizamos el state de setconsultAPI el cual guardaba true eso es que hacía la consulta a la BBDD y actualizaba los registros
     const deleteAppointment = async (id) => {
         // console.log(id);
-    
-        axiosClient.delete(`/patients/${id}`)
-            .then(res => {
-                props.setconsultAPI(true)
-                props.history.push('/') 
 
-            })
-            .catch (e => {
-                console.log(e);
-                
-            })
+        //Alerta para confirmar el elminimado de la cita (sweetalert2)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!',
+            cancelButtonText: 'Cancel'
+        }).then((result) => {
+            if (result.isConfirmed) {
+
+                //Alerta de eliminado
+                Swal.fire(
+                    'Deleted!',
+                    'Your file has been deleted.',
+                    'success'
+                )
+
+                //Eliminado de la BBDD
+                axiosClient.delete(`/patients/${id}`)
+                    .then(res => {
+                        props.setconsultAPI(true)
+                        props.history.push('/')
+
+                    })
+                    .catch(e => {
+                        console.log(e);
+
+                    })
+            }
+        })
 
     }
 
@@ -60,7 +84,7 @@ const AppointmentDetail = (props) => {
                                 </div>
                                 <div>
                                     {/*al onClick no le puedo pasar directamente una función ya que se estaría ejecutando tdo el tiempo y nosotros solo queremos que se ejecutre cuando lo clicke el ususario, además, como estamos eliminando, hay que pasarle un id, por tanto hay que hacer un arrow function y pasar por parámetro el id*/}
-                                    <button type="button"  className="delete" onClick={() => deleteAppointment(props.oneappointment._id)}>&#x1f5d1;</button>
+                                    <button type="button" className="delete" onClick={() => deleteAppointment(props.oneappointment._id)}>&#x1f5d1;</button>
                                 </div>
                             </div>
                         </div>
